cfg relation_type_weights is now a double for more flexibility

+++ engine/cfg.h
@@ -25,7 +25,7 @@ CORONA_CFG(double, "%.4f", probability_asymptomatic)
 CORONA_CFG(double, "%.4f", probability_mild)
 CORONA_CFG(double, "%.4f", probability_critical)
 
-CORONA_CFG_VECTOR(uint32_t, "%u", relation_type, relation_type_weights, NUMBER_OF_RELATIONS)
+CORONA_CFG_VECTOR(double, "%.4f", relation_type, relation_type_weights, NUMBER_OF_RELATIONS)
 
 CORONA_CFG_VECTOR(double, "%.4f", infected_state, r0_factor_per_group, NUMBER_OF_INFECTED_STATES)
